@@ -7,10 +7,10 @@
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
     "clarifai": "^2.9.1",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-particles-js": "^3.4.0",
-    "react-scripts": "3.4.4",
+    "react-scripts": "5.0.1",
     "react-tilt": "^0.1.4",
     "tachyons": "^4.12.0"
   },
