@@ -1,39 +1,33 @@
-import React from "react"; //importa la librería de React
+import React from "react";
 
 class SignIn extends React.Component {
-  // define "SignIn" como una instancia de "Component" definido por react
   constructor(props) {
-    //la forma de pasarle métodos y variables a este objeto es a través del parámetro "props" del constructor
     super(props);
     this.state = {
-      //creación de dos variables que toman los datos necesarios para que el usuario entre en la app
       signInEmail: "",
       signInPassword: "",
     };
   }
-  // función para escribir la variable sigInEmail de acuerdo al input del usuario
+
   onEmailChange = (event) => {
     this.setState({ signInEmail: event.target.value });
   };
 
-  // función para escribir la variable sigInPassword de acuerdo al input del usuario
   onPasswordChange = (event) => {
     this.setState({ signInPassword: event.target.value });
   };
 
   //Función que toma los inputs de los usuarios y los compara con los usuarios registrados en la base de datos
   onSubmitSignIn = () => {
-    fetch("http://localhost:3002/signin", {
-      //HTTP request, mandando la información puesta por el usuario
-      method: "post", //método del request
-      headers: { "Content-Type": "application/json" }, //header que identifica el tipo de body
+    fetch("http://localhost:8080/signin", {
+      method: "post",
+      headers: { "Content-Type": "application/json" },
       body: JSON.stringify({
-        //en el body se manda en formato JSON el objeto javascript
         email: this.state.signInEmail,
         password: this.state.signInPassword,
       }),
     })
-      .then((response) => response.json()) //a la respuesta se la pasa de json a javascript
+      .then((response) => response.json())
       .then((user) => {
         if (user.id) {
           // si existe un id significa que en la respuesta se mandó un usuario y por lo canto se puede ejecutar las funciones de cargar ese usuario en el root y cambiarla ruta así se muestra entonces esa página
@@ -77,7 +71,7 @@ class SignIn extends React.Component {
             </fieldset>
             <div className="">
               <input
-                onClick={this.onSubmitSignIn} //Arrow function so the function doesnt get triggered when the app runs; i.e., hacerla un callback (si es que entiendo bien)
+                onClick={this.onSubmitSignIn}
                 className="b ph3 pv2 input-reset ba b--black bg-transparent grow pointer f6 dib"
                 type="submit"
                 value="Sign in"
